fix(?) something
so part of the reason twelveth auto chall is so hard is because costs increase past 1.8e308. that's because instead of increasing past 2^2^12 it increases past 2^2^10. however, that makes no sense. outside the challenge it should increase past 2^2^10 because it says costs increase faster post-infinity, not post-infinity^4.

idea: break infinity upgrade that raises normal_corruption_start. when it's first raised, normal_corruption_start is lowered in challenge 12 to 2^2^11. when it's 2^2^12 it's lowered in chall 12 to 2^2^10. when it's more than 2^2^12 it's the same in chall 12.